Refactor: 클래스 구조 개선
