+++ src/main/java/support/instance/DefaultInstanceManager.java
@@ -1,4 +1,4 @@
-package support;
+package support.instance;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
