+++ src/main/java/support/instance/InstanceInitializer.java
@@ -1,4 +1,4 @@
-package support;
+package support.instance;
 
 import support.annotation.Container;
 import utils.ClassListener;
