+++ src/main/java/support/web/ControllerMethodStruct.java
@@ -1,4 +1,4 @@
-package support;
+package support.web;
 
 import java.lang.reflect.Method;
