+++ src/main/java/support/web/ViewResolver.java
@@ -1,0 +1,4 @@
+package support.web;
+
+public class ViewResolver {
+}
