+++ src/test/java/support/web/ControllerResolverTest.java
@@ -1,4 +1,4 @@
-package support;
+package support.web;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
